requirements.txt
# Business Card Processing API - Dependencies
# All free, no paid APIs required

# Web Framework
Flask>=2.3.0
Werkzeug>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# OCR Engine (Free, local processing)
easyocr>=1.7.0

# Vision Language Model (Gemini - nearly free fallback)
google-genai>=1.0.0

# Image Processing
Pillow>=10.4.0
opencv-python-headless>=4.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP Requests (for free API enrichment)
requests>=2.31.0

# Environment & Configuration
python-dotenv>=1.0.0

# Utilities
phonenumbers>=8.13.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Production Server (optional)
gunicorn>=21.0.0